code prettyfing and the addtion of the end game stage
