+++ lib/Pages/GameForm/widgets/Summery/SummeryPage.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:scouting_app_2/Pages/GameForm/widgets/Summery/EndStageWidget.dart';
 import 'package:scouting_app_2/Pages/GameForm/widgets/Summery/MidGameWidget.dart';
 import 'package:scouting_app_2/models/Match/summery/ScoutingMatchSummery.dart';
 
@@ -13,7 +14,7 @@ class SummeryPage extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
         body: PageView(
-          scrollDirection: Axis.vertical,
+      scrollDirection: Axis.vertical,
       children: [
         MidStageWidget(
           title: "Autonumos",
@@ -23,47 +24,10 @@ class SummeryPage extends StatelessWidget {
           title: "Teleop",
           data: summery.matchData.auto,
         ),
+        EndStageWidget(summery.matchData.endGame),
       ],
     ));
   }
 }
 
-class MidStageWidget extends StatelessWidget {
-  final String title;
-  final MidGameDataSummery data;
-  MidStageWidget({this.title, this.data});
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: [
-        Row(
-          mainAxisAlignment: MainAxisAlignment.start,
-          children: [
-            Padding(
-              child: Text("$title:"),
-              padding: EdgeInsets.fromLTRB(15, 0, 0, 0),
-            )
-          ],
-        ),
-        SizedBox(
-          height: 2,
-        ),
-        Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Container(
-              color: Colors.grey,
-              height: 1,
-              width: MediaQuery.of(context).size.width * 0.9,
-            )
-          ],
-        ),
-        SizedBox(
-          height: 1,
-        ),
-        MidGameStageWidget(data),
-      ],
-    );
-  }
-}
+
